prevent deleting already deleted auth accounts in deleteUser action

+++ src/lib/actions/admin/users-actions.ts
@@ -152,7 +152,7 @@ export const updateVerified = async (userId: string, isVerified: boolean, metada
     return { success: true, message: `Researcher ${action} successfully.` };
 };
 
-export const alertUser  = async (userId: string, message: string) => {
+export const alertUser = async (userId: string, message: string) => {
     const supabase = createAdminClient();
 
     const notification = {
@@ -178,13 +178,13 @@ export const deleteUsers = async (ids: string[], usernames?: string[], deletionT
 
     for (let i = 0; i < ids.length; i++) {
         // Update deleted_at in database
-        const { error: dbError } = deletionType === "soft" 
-        ? await supabase.from('users').update({
-            deleted_at: new Date().toISOString(),
-            username: `${usernames![i]} (deleted)`, // mark the username as deleted to allow reusing it
-            profile_picture: null
-        }).eq('id', ids[i])
-        : await supabase.from('users').delete().eq('id', ids[i]);
+        const { error: dbError, data } = deletionType === "soft"
+            ? await supabase.from('users').update({
+                deleted_at: new Date().toISOString(),
+                username: `${usernames![i]} (deleted)`, // mark the username as deleted to allow reusing it
+                profile_picture: null
+            }).eq('id', ids[i])
+            : await supabase.from('users').delete().eq('id', ids[i]).select("deleted_at").single();
 
         if (dbError) {
             console.log('Error deleting user from db:', dbError);
@@ -192,12 +192,14 @@ export const deleteUsers = async (ids: string[], usernames?: string[], deletionT
             continue;
         }
 
-        // Delete the Supabase Auth account
-        const { error: authError } = await supabase.auth.admin.deleteUser(ids[i]);
-        if (authError) {
-            console.log('Error deleting Supabase Auth account:', authError);
-            failed++;
-            continue;
+        // Delete the Supabase Auth account if not already deleted before
+        if (!data?.deleted_at) {
+            const { error: authError } = await supabase.auth.admin.deleteUser(ids[i]);
+            if (authError) {
+                console.log('Error deleting Supabase Auth account:', authError);
+                failed++;
+                continue;
+            }
         }
 
         // Remove profile picture in Supabase Storage to save space since it is not needed for auditing
